Print system information for each node and use separate output files for each federate
